Update draw_text.rb
This should fix https://github.com/kojix2/LibUI/issues/74 - I looked carefully and I think two lines are simply duplicates. I am not sure whether this is by intent or not, perhaps it is to avoid a pointer issue, but I checked the changed example locally and it worked after my change, so I assume these two lines are not needed, as they are the same as the two lines above. I am making this detailed so that kojix2 understands the small change made here. :)